Finished Responsive Navbar with HTML, CSS and JS
This responsive Navbar was made by some html, css and js. The Javascript on the click of the burger icon, adds a class to the navbar-links and removes it by clicking on it again. It has an eventListener to the window too that gets triggered by resizing the viewport width and, removes the show-links class from the navbar-links if it's already a class of the navbar-links and if the viewport width is greater equal of 800